+++ Pandora/Pattern/Morphism.hs
@@ -3,6 +3,8 @@ module Pandora.Pattern.Morphism (module Exports, Opposite) where
 import Pandora.Pattern.Morphism.Trip as Exports
 import Pandora.Pattern.Morphism.Flip as Exports
 import Pandora.Pattern.Morphism.Tensor as Exports
+import Pandora.Pattern.Morphism.Tensor as Exports
+import Pandora.Pattern.Morphism.Lifted as Exports
 import Pandora.Pattern.Morphism.Kleisli as Exports
 import Pandora.Pattern.Morphism.Straight as Exports