+++ Pandora/Pattern/Morphism/Lifted.hs
@@ -1,0 +1,3 @@
+module Pandora.Pattern.Morphism.Lifted where
+
+newtype Lifted t (v :: * -> * -> *) source target = Lifted (v (t source) (t target))